changed: split VideoUtils in two parts
GUI and non-GUI dependent code
