+++ xbmc/favourites/GUIWindowFavourites.cpp
@@ -23,6 +23,7 @@
 #include "utils/StringUtils.h"
 #include "utils/guilib/GUIContentUtils.h"
 #include "video/VideoUtils.h"
+#include "video/guilib/VideoGUIUtils.h"
 #include "video/guilib/VideoPlayActionProcessor.h"
 #include "video/guilib/VideoSelectActionProcessor.h"
