+++ xbmc/video/guilib/VideoSelectActionProcessor.cpp
@@ -21,7 +21,7 @@
 #include "utils/StringUtils.h"
 #include "utils/Variant.h"
 #include "video/VideoInfoTag.h"
-#include "video/VideoUtils.h"
+#include "video/guilib/VideoGUIUtils.h"
 
 using namespace VIDEO::GUILIB;
